Fixed #2
We didn't go with the script I'd used initially. Greg suggested making
it python and I'd wanted to clean up the interface to make it consistent
to what I wanted the Tootsuite tools to look like, so I've made a much
simpler and fairly streamlined approach. I'll want to confirm it against
the work I did in January, but tentatively it looks solid.